# crates/mica-apid/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "apid",
    "description": "The appliance management API served over HTTPS on the device.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/provisioning/status": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Read what a provisioning document did to this device.",
        "description": "Two settings reads plus the public baked configuration: the record micad\nwrote when it last met a document, and whether an administrator credential\nexists. Nothing is observed from a medium at request time — the media are staged and read\nbefore anything is listening, so a request cannot make a device look at a\nstick.",
        "operationId": "api_v1_provisioning_status",
        "responses": {
          "200": {
            "description": "Import history and claim state, the public baked manifest with SHA-256 digests by baked file path, and the operator and effective readings of `update.source`, `update.channel` and `update.policy` beside the baked `fleet` pair. The secret-bearing import document is never returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningStatus"
                }
              }
            }
          },
          "401": {
            "description": "No stored bearer token or authenticated browser session (`not_authenticated`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "A method this route does not serve (`method_not_allowed`); carries `Allow`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "micad failed to answer (`micad_failed`), the baked configuration could not be read (`baked_configuration_unavailable`), or `/mica/config/updates.json` did not read, parse or validate (`configuration_unavailable`) — which is refused rather than answered with the baked value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The call to micad could not be made (`micad_unreachable`); carries `Retry-After`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "The bounded call to micad timed out (`micad_timeout`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/session": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Report setup and browser authentication state without redirecting.",
        "operationId": "api_v1_session_status",
        "responses": {
          "200": {
            "description": "Setup and browser authentication state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            }
          },
          "500": {
            "description": "The access settings could not be read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "micad is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Authenticate a browser and create its HttpOnly session cookie.",
        "operationId": "api_v1_session_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The browser session was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            }
          },
          "400": {
            "description": "The body is not JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "The password is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The device still requires first-run setup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "The JSON body has the wrong shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Login attempts are temporarily throttled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session"
        ],
        "summary": "Revoke the acting browser session.",
        "operationId": "api_v1_session_delete",
        "responses": {
          "204": {
            "description": "The browser session was revoked"
          },
          "401": {
            "description": "No API credential was supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The browser CSRF token is absent or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "The one shape every failure under `/api/` takes.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "description": "The API error envelope payload.",
        "required": [
          "code",
          "message",
          "source"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string"
          }
        }
      },
...
      "ProvisioningStatus": {
        "type": "object",
        "description": "What a provisioning document did to this device.",
        "required": [
          "unclaimed",
          "baked",
          "bakedDigests",
          "operator",
          "effective"
        ],
        "properties": {
          "baked": {
            "description": "The public configuration baked into the verity root, as read."
          },
          "bakedDigests": {
            "type": "object",
            "description": "SHA-256 by relative path under `/usr/share/mica/meta/`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "documentDigest": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical digest of the document last applied; `null` with the version."
          },
          "documentVersion": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "`version` of the document last applied; `null` when none ever was.",
            "minimum": 0
          },
          "effective": {
            "description": "What this device actually runs on after precedence:\n`update.source`, `update.channel`, `update.policy`, and `fleet.url` /\n`fleet.enabled`.\n\nThe fleet pair is the **baked** value and is not a placeholder for a\nresolution not yet written: `/mica/config/`'s fleet document\ndoes not exist, so nothing on the device reads an\noperator layer for it either. When that slice lands it extends the\nlibrary's resolver; it does not add a second one here."
          },
          "lastImport": {
            "description": "The last import ATTEMPT: `source` (`boot` or `media`), `outcome`\n(`applied`, `unchanged` or `rejected`), `reason` for a rejection, and\n`at`, the device clock's reading when it happened. `null` when no\nmedium has ever been offered to this device."
          },
          "operator": {
            "description": "What the operator wrote in `/mica/config/updates.json`, projected onto\n`update.source`, `update.channel` and `update.policy` and nothing else.\n\nA field the operator did not write is **absent**; one they wrote as\n`null` is **`null`**. Both resolve to the baked default and they are\nstill two different statements about the document, so the reading is\npreserved rather than flattened. `{}` when the document overrides\nnothing, which is a factory-reset device."
          },
          "unclaimed": {
            "type": "boolean",
            "description": "Whether the device still has no administrator credential.\n\nThe same question `GET /api/v1/session` answers with `state: \"setup\"`,\nasked here because it is what decides whether a document offered on a\nmedium would be applied at all: a claimed device refuses one."
          }
        }
      },
...
      "SessionLoginRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
      "SessionStatus": {
        "type": "object",
        "description": "The browser's current authentication state.",
        "required": [
          "state"
        ],
        "properties": {
          "csrfToken": {
            "type": [
              "string",
              "null"
            ],
            "description": "Returned only for an authenticated browser session."
          },
          "state": {
            "type": "string",
            "description": "`setup`, `unauthenticated`, or `authenticated`."
          }
        }
      },
...
    }
  }
}